+++ src/redturtle/patterns/slider/static/pattern.js
@@ -1,27 +1,27 @@
-define('redturtle-patterns-slider', [
-  'jquery',
-  'pat-base',
-  'slick.min',
-], function ($, Base) {
-  'use strict';
+define("redturtle-patterns-slider", [
+  "jquery",
+  "pat-base",
+  "collective-slick-js"
+], function($, Base) {
+  "use strict";
   var Slider = Base.extend({
-    name: 'slider',
-    trigger: '.pat-slider',
-    parser: 'mockup',
+    name: "slider",
+    trigger: ".pat-slider",
+    parser: "mockup",
     defaults: {
       dots: true,
       slidesToShow: 4,
       slidesToScroll: 4,
       arrows: true,
-      lazyLoad: 'ondemand',
+      lazyLoad: "ondemand"
     },
-    init: function () {
+    init: function() {
       var _this = this;
       _this.$el.slick(_this.options);
 
-      var event = new CustomEvent('patSliderInit', { detail: _this.$el[0] });
+      var event = new CustomEvent("patSliderInit", { detail: _this.$el[0] });
       document.dispatchEvent(event);
-    },
+    }
   });
 
   return Slider;